+++ src/xyz/oihana/schema/constants/Oihana.php
@@ -2,6 +2,7 @@
 
 namespace xyz\oihana\schema\constants;
 
+use xyz\oihana\schema\constants\traits\AuthTrait;
 use xyz\oihana\schema\constants\traits\LogTrait;
 use xyz\oihana\schema\constants\traits\PaginationTrait;
 
@@ -12,6 +13,7 @@ class Oihana
      */
     public const string SCHEMA = 'https://schema.oihana.xyz';
 
-    use LogTrait,
+    use AuthTrait ,
+        LogTrait,
         PaginationTrait ;
 }