+++ src/xyz/oihana/schema/constants/traits/auth/PermissionTrait.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace xyz\oihana\schema\constants\traits\auth;
+
+/**
+ * The enumeration of all Permission properties.
+ *
+ * @package xyz\oihana\schema\constants\traits
+ * @author  Marc Alcaraz (ekameleon)
+ * @since   1.0.2
+ */
+trait PermissionTrait
+{
+    const string ACTION  = 'action'  ;
+    const string DOMAIN  = 'domain'  ;
+    const string SUBJECT = 'subject' ;
+}