+++ src/xyz/oihana/schema/constants/traits/auth/RoleTrait.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace xyz\oihana\schema\constants\traits\auth;
+
+/**
+ * The enumeration of all Roles properties.
+ *
+ * @package xyz\oihana\schema\constants\traits
+ * @author  Marc Alcaraz (ekameleon)
+ * @since   1.0.2
+ */
+trait RoleTrait
+{
+    const string PERMISSIONS       = 'permissions' ;
+    const string PERMISSIONS_COUNT = 'permissionsCount' ;
+    const string USERS             = 'users' ;
+    const string USERS_COUNT       = 'usersCount' ;
+}